requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.txt
#
amqp==5.1.1
billiard==3.6.4.0
Brotli==1.0.9
CacheControl==0.12.11
cachetools==5.0.0
celery==5.2.6
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
firebase-admin==5.2.0
Flask==2.1.2
Flask-Compress==1.12
Flask-Cors==3.0.10
google-api-core==2.7.2
google-api-python-client==2.46.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-cloud-core==2.3.0
google-cloud-firestore==2.4.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpcio==1.44.0
grpcio-status==1.44.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kombu==5.2.4
ldap3==2.9.1
lxml==4.8.0
MarkupSafe==2.1.1
msgpack==1.0.3
numpy==1.22.3
nyoka==5.2.0
pandas==1.4.2
prompt-toolkit==3.0.29
proto-plus==1.20.3
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.8
pyjwt==2.4.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
rsa==4.8
scikit-learn==1.6.1
scipy==1.8.0
simplejson==3.17.6
six==1.16.0
threadpoolctl==3.1.0
tinydb==4.7.0
uritemplate==4.1.1
urllib3==1.26.9
uwsgi==2.0.20
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.1.2
xgboost==0.90
zipp==3.8.0
